Work on tomas test

+++ kattis/segmentintersection/main.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 using namespace popup;
 
-#define rep(i, a, b) for(int i = a; i < (b); ++i)
+#define rep(i, a, b) for(int i = a; i < (int)(b); ++i)
 #define trav(a, x) for(auto& a : x)
 #define all(x) x.begin(), x.end()
 #define sz(x) (int)(x).size()
